requirements.txt
ttkbootstrap>=1.10.1
Pillow>=10.0.0
tkinterdnd2>=0.3.0 